Lots of changes! Most of the abstract interpretation works now
This commit is a bit of a mess. I've been hacking wildly for too
long. But progress is coming along nicely.

I made a long detour looking into LogicT but it is way overkill for
what I'm doing here. MonadPlus is enough.

I'm very displeased with how I've handled multiple state monads.
I've created two new monads with their corresponding classes.
This really doesn't scale! One way to deal with the problem would
be to have a way to generate the classes. But I should ask around
about better ways to deal with this.