Various minor improvements
I've tried to get symbolic execution to work but I don't see how to do that with the current setup.
The mixes values and expressions freely in symbolic execution which doesn't work well
for the way I've set up the types.

I've also made some improvements to the GC code. But ultimately I don't understand
how to implement all of it. The paper is very sparse with details around GC and there's not
enough detail there for me to figure out how it's supposed to work